3a) Wrong diagnostic mentioning -T option (should be -D!) corrected in parsecl.c and in comments in dieharder.h
Was:
$ dieharder -g 201 -f xaf -p 1 -d 4 -D histogr
Invalid -T histogr option.

3b) nonexisting tests leads to false diagnostic

Was:
$ dieharder -g 201 -f bits.01 -D 18558  -D pvalues -D description -t 1000000 -d 18 -p 1
Segmentation fault: 11

Corrected: the test for the correctness of the test number and/or name is added to dieharder/parsecl.c, and in dieharder/dieharder.c the calls to  dieharder_rng_types() and dieharder_test_types() are moved (as it was recommended in the comments!) before the call to parsecl(). The conversion from test names to test numbers is also moved here from run_test.c, so correct error message is possible.

3c) These changes also makes call to dieharder_rng_types in dieharder/list_rngs.c, line 37, unnecessary.

Note also that add_un_rngs() in the next line 38 makes no sense here, it should be done in the main loop when the testing happens -- and then this line would be also not necessary. [By the way, the instructiond in dieharder/add_ui_rngs.c are contradictory: an example is mentioned by it is not present. No changes.]

3d) List of all generators (dieharder -g -1) is printed with "Test Name" (in the header, three times) instead of "Generator Name", see dieharder/list_rngs.c. [Now corrected]